Change LoggerConfiguration.LogLevel to be of type Microsoft.Extensions.Logging.LogLevel
- Change LoggerConfiguration.LogLevel to be of type Microsoft.Extensions.Logging.LogLevel
- Use string.Equals(str1, str2, StringComparison. InvariantCultureIgnoreCase) to compare strings, ignoring case.

+++ LoRaEngine/modules/LoRaWanNetworkSrvModule/Logger/Logger.cs
@@ -66,7 +66,7 @@ namespace LoRaWan
 
         public static void Log(string deviceId, string message, LogLevel logLevel)
         {
-            if ((int)logLevel >= configuration.LogLevel)
+            if ((int)logLevel >= (int)configuration.LogLevel)
             {
                 var msg = message;
